--- scripts/docs-collator/ComponentRepositoryAbstract.py
@@ -1,3 +1,0 @@
-
-class ComponentRepositoryAbstract:
-    pass
