+++ scripts/docs-collator/component/renderer/templates/components/multiple/index_category.json
