+++ scripts/docs-collator/component/renderer/templates/components/multiple/readme.md
