+++ scripts/docs-collator/component/renderer/templates/components/multiple/terraform-docs.yml
