+++ scripts/docs-collator/component/renderer/templates/components/single/index_category.json
