+++ scripts/docs-collator/component/renderer/templates/components/single/readme.md
