+++ scripts/docs-collator/component/renderer/templates/components/single/terraform-docs.yml
